eval_pipeline/requirements.txt
functions-framework==3.*
pandas
google-cloud-aiplatform
google-cloud-aiplatform[evaluation]
google-cloud-bigquery
bigframes
pandas-io
pandas-gbq